Move common test members to auth_test.go
